backend/requirements.txt
Django==5.0.6
django-environ==0.11.2
psycopg==3.1.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
Pillow==10.4.0
django-admin-thumbnails==0.2.9
django-cors-headers==4.0.0
drf-spectacular==0.29.0